+++ lib/index.d.ts
@@ -489,6 +489,12 @@ declare namespace Realm.Sync {
         validateCallback?: SSLVerifyCallback;
     }
 
+    const enum ClientResyncMode {
+        Discard = 'discard',
+        Manual = 'manual',
+        Recover = 'recover'
+    }
+
     interface SyncConfiguration {
         user: User;
         url: string;
@@ -505,9 +511,10 @@ declare namespace Realm.Sync {
         _disableQueryBasedSyncUrlChecks?: boolean;
         _sessionStopPolicy?: SessionStopPolicy;
         custom_http_headers?: { [header: string]: string };
-        customQueryBasedSyncIdentifier?: string,
-        newRealmFileBehavior?: OpenRealmBehaviorConfiguration
-        existingRealmFileBehavior?: OpenRealmBehaviorConfiguration
+        customQueryBasedSyncIdentifier?: string;
+        newRealmFileBehavior?: OpenRealmBehaviorConfiguration;
+        existingRealmFileBehavior?: OpenRealmBehaviorConfiguration;
+        clientResyncMode?: ClientResyncMode;
     }
 
     interface OpenRealmBehaviorConfiguration {