+++ src/js_sync.hpp
@@ -1254,8 +1254,25 @@ void SyncClass<T>::populate_sync_config(ContextType ctx, ObjectType realm_constr
         }
 #endif
 
-    // FIXME: remove once Client Resync is implemented
-    config.sync_config->client_resync_mode = realm::ClientResyncMode::Manual;
+        // default for query-based sync is manual and recover for full sync
+        ClientResyncMode clientResyncMode = (config.sync_config->is_partial) ? realm::ClientResyncMode::Manual : realm::ClientResyncMode::Recover;
+        ValueType client_resync_mode_temp = Object::get_property(ctx, sync_config_object, "clientResyncMode");
+        if (!Value::is_undefined(ctx, client_resync_mode_temp)) {
+            std::string client_resync_mode = std::string(Value::validated_to_string(ctx, client_resync_mode_temp, "client_resync_mode"));
+            if (client_resync_mode == std::string("recover")) {
+                clientResyncMode = realm::ClientResyncMode::Recover;
+            } else if (client_resync_mode == std::string("manual")) {
+                clientResyncMode = realm::ClientResyncMode::Manual;
+            } else if (client_resync_mode == std::string("discard")) {
+                clientResyncMode = realm::ClientResyncMode::DiscardLocal;
+            } else {
+                throw std::invalid_argument("Unknown argument for clientResyncMode: " + client_resync_mode);
+            }
+        }
+        if (config.sync_config->is_partial && clientResyncMode != realm::ClientResyncMode::Manual) {
+            throw std::invalid_argument("Only 'manual' resync mode is supported for query-based sync.");
+        }
+        config.sync_config->client_resync_mode = clientResyncMode;
     }
 }